Chatbot_and_Feedback/requirements.txt
Django>=4.2
psycopg2-binary
requests
python-dotenv